llama-vision-chat/requirements.txt
requests
torch
Pillow
transformers